tableWithOdp: remove unused code/files, move some temporarely under contentCheck
