--- webapp/app/assessment/fra/components/tableWithOdp/components/generateValues/index.js
@@ -1,1 +1,0 @@
-export { default } from './generateValues'
