--- webapp/app/assessment/fra/components/tableWithOdp/components/table/index.js
@@ -1,1 +1,0 @@
-export { default } from './table'
