--- webapp/app/assessment/fra/components/tableWithOdp/components/tableWithOdp.less
@@ -1,71 +1,0 @@
-@import (reference) "~@webapp/main/styles/config";
-@import (reference) "~@webapp/main/styles/fraTable";
-
-.table-with-odp {
-  .odp-header-cell {
-    .mixin-table-cell(700, center, 0);
-    background-color: fade(@ui-accent, 10%);
-
-    .link {
-      padding: 8px;
-      text-decoration: underline;
-    }
-  }
-
-  .odp-value-cell {
-    background-color: fade(@ui-accent, 5%);
-  }
-
-  .odp-value-cell-total {
-    .mixin-table-cell(700, right, 10px, true);
-    background-color: fade(@ui-accent, 5%);
-  }
-
-  .validation-error {
-    background-color: fade(@ui-destructive, 10%);
-  }
-}
-
-// Generate
-
-.table-with-odp__generate-control {
-  .select-s {
-    width: 180px;
-  }
-
-  .btn-s {
-    margin-left: 8px;
-  }
-}
-
-.table-with-odp__generate-inputs-table {
-  border-collapse: collapse;
-  font-size: @font-s;
-  margin-top: 8px;
-}
-
-.table-with-odp__generate-inputs-row-selection-cell {
-  width: 20px;
-  height: 32px;
-}
-
-.table-with-odp__generate-input-header {
-  padding: 2px 8px 2px 2px;
-  width: 164px;
-}
-
-.table-with-odp__generate-input-cell {
-  padding: 2px 0 2px 4px;
-  width: 62px;
-
-  .text-input-s {
-    width: 100%;
-  }
-}
-
-.table-with-odp__generate-comment-cell {
-  font-size: @font-xxs;
-  padding: 0 4px;
-  color: @text-mute;
-}
-
