--- webapp/app/assessment/fra/components/tableWithOdp/index.js
@@ -1,1 +1,0 @@
-export { default } from './components/tableWithOdp'
